move rarity graph to its own file
